TIDY MOSULES AND FIX ISSUE WITH ADDING TO DOS BEING BRIEF
Moved the contents of pageload.js to index.js and moved lists.js back
to addToDo.js to stop having too many/unnecessary modules.

When adding to do's to list the page refreshed and any messages in the
console or actions on the page only appeared briefly. This was due to having
the form to add a to do in a form elemnent with no action given. The fix used
was to change the form element to a div.