ADD EXPAND ICON, RESET FORM FIELDS, REMOVE PREVIOUS CONTENT
Added an expand icon to be used to click on and exopand view of to do
or project.

The form fields for adding a to do now reset to being empty after
clicking to add a to do.

If navigating from one area to another using the nav bar the previous areas
content is now removed from main before displaying the one navigated to.